refactor: ScheduleDrawer 구조 개선 및 기본 탭 페이지 추가
