+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/CategoryPicker.tsx
@@ -1,10 +1,23 @@
 import { SyntheticEvent, useState } from "react";
 import { Tab, Tabs } from "@mui/material";
+import ExpenditureCategoryPage from "./pages/ExpenditureCategoryPage";
+import IncomeCategoryPage from "./pages/IncomeCategoryPage";
 
 function CategoryPicker() {
   const [activeTabIndex, setActiveTabIndex] = useState(0);
   const handleChange = (event: SyntheticEvent, newValue: number) => {
     setActiveTabIndex(newValue);
+  };
+
+  const content = (activeTabIndex: number) => {
+    switch (activeTabIndex) {
+      case 0:
+        return <ExpenditureCategoryPage />;
+      case 1:
+        return <IncomeCategoryPage />;
+      default:
+        return <div>지출</div>;
+    }
   };
 
   return (
@@ -19,6 +32,7 @@ function CategoryPicker() {
         <Tab label="지출" />
         <Tab label="수입" />
       </Tabs>
+      {content(activeTabIndex)}
     </>
   );
 }
