+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/components/CategorySideTabButton.stories.tsx
