+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/components/CategorySideTabButton.tsx
