+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/pages/ExpenditureCategoryPage/ExpenditureCategoryPage.tsx
@@ -1,0 +1,5 @@
+function ExpenditureCategoryPage() {
+  return <div>ExpenditureCategoryPage</div>;
+}
+
+export default ExpenditureCategoryPage;
