+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/pages/IncomeCategoryPage/IncomeCategoryPage.tsx
@@ -1,0 +1,5 @@
+function IncomeCategoryPage() {
+  return <div>IncomeCategoryPage</div>;
+}
+
+export default IncomeCategoryPage;
