+++ src/containers/home/ScheduleDrawer/inputs/CategoryPicker/pages/IncomeCategoryPage/index.ts
@@ -1,0 +1,3 @@
+import IncomeCategoryPage from "./IncomeCategoryPage";
+
+export default IncomeCategoryPage;
